pyproject.toml
[tool.poetry]
name = "homework"
version = "0.1.0"
description = ""
authors = ["Piotr Rogulski <rivinek@gmail.com>", "Artem Korzhenevsky <artkorzh1@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85"
SQLAlchemy = "~1.4.25"
greenlet = "^1.1.3"
uvicorn = "^0.19.0"
asyncpg = "^0.27.0"
alembic = "^1.8"

[tool.poetry.dev-dependencies]
black = "^22.10"
click = "^8.1.3"
pytest = "^7.2"
httpx = "^0.23.0"
pytest-asyncio = "^0.20.1"
PyHamcrest = "^2.0.4"

[build-system]
requires = ["poetry-core>=1.3.2"]
build-backend = "poetry.core.masonry.api"
